relations

+++ src/idea/idea.service.ts
@@ -1,8 +1,9 @@
+import { UserEntity } from './../user/user.entity';
 import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from './idea.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { IdeaDTO } from './idea.dto';
+import { IdeaDTO, IdeaRO } from './idea.dto';
 
 @Injectable()
 export class IdeaService {
@@ -10,41 +11,56 @@ export class IdeaService {
   constructor(
     @InjectRepository(IdeaEntity)
     private ideaRepository: Repository<IdeaEntity>,
+    @InjectRepository(UserEntity)
+    private userRepository: Repository<UserEntity>
   ) {}
 
-  async showAll() {
-    return await this.ideaRepository.find();
+  private toResponseObject(idea: IdeaEntity): IdeaRO {
+    return {...idea, author: idea.author.toResponseObject(false)};
   }
 
-  async create(data: IdeaDTO) {
-    const idea = await this.ideaRepository.create(data);
+  private ensureOwnership(idea, userId) {
+    if(idea.author.id !== userId) {
+      throw new HttpException('Incorrect user', HttpStatus.UNAUTHORIZED);
+    }
+  }
+  async showAll(): Promise<IdeaRO[]> {
+    const ideas = await this.ideaRepository.find({relations: ['author']});
+    return ideas.map(idea => this.toResponseObject(idea));
+  }
+
+  async create(userId: string, data: IdeaDTO): Promise<IdeaRO> {
+    const author = await this.userRepository.findOne({id: userId});
+    const idea = await this.ideaRepository.create({...data, author});
     await this.ideaRepository.save(idea);
-    return idea;
+    return this.toResponseObject(idea);
   }
 
-  async show(id: string) {
-    const idea = await this.ideaRepository.findOne({ id }); 
+  async show(id: string): Promise<IdeaRO> {
+    const idea = await this.ideaRepository.findOne({ where: {id}, relations: ['author'] }); 
     if (!idea) {
       throw new HttpException('Not found', HttpStatus.NOT_FOUND);
     }
-    return idea;
+    return this.toResponseObject(idea);
   }
 
-  async update(id: string, data: Partial<IdeaDTO>) {
-    const idea = await this.ideaRepository.findOne({ id }); 
+  async update(id: string, userId: string,data: Partial<IdeaDTO>): Promise<IdeaRO> {
+    const idea = await this.ideaRepository.findOne({ where: {id}, relations: ['author'] }); 
     if (!idea) {
       throw new HttpException('Not found', HttpStatus.NOT_FOUND);
     }
+    this.ensureOwnership(idea, userId);
     await this.ideaRepository.update({ id }, data);
-    return {...idea, ...data};
+    return this.toResponseObject({...idea, ...data});
   }
 
-  async delete(id: string) {
-    const idea = await this.ideaRepository.findOne({ id }); 
+  async delete(id: string, userId: string): Promise<IdeaRO> {
+    const idea = await this.ideaRepository.findOne({ where:{id}, relations: ['author'] }); 
     if (!idea) {
       throw new HttpException('Not found', HttpStatus.NOT_FOUND);
     }
+    this.ensureOwnership(idea, userId);
     await this.ideaRepository.delete({ id });
-    return idea;
+    return this.toResponseObject(idea);
   }
 }
